refactor: rename serverSettings utils
